Add basic "explorer" GUI for the Mandelbrot set (#66)
Uses the [`winit`](https://docs.rs/winit/latest/winit/index.html) and [`pixels`](https://docs.rs/pixels/latest/pixels/) libraries to implement the back end of a basic GUI that can be used to explore the Mandelbrot set. Most of the implementation is agnostic to the specific fractal, so it will be easy to add support for the DDP fractal (and eventually others).

Also updates some of the data types used in the final images to use `f32` instead of `f64`. Continue using `f64` for core calculations.

+++ src/core/chaos_game.rs
@@ -127,7 +127,7 @@ where
         pixel.apply2(&background_color, |background: u8, pixel: u8| -> u8 {
             ((background as f32) * weight_background + (pixel as f32) * weight_pixel) as u8
         });
-        histogram.insert(weight_background as f64);
+        histogram.insert(weight_background);
     }
     timer.antialiasing_post_process = elapsed_and_reset(&mut stopwatch);
